extract ros2_reader_base, share common logic between readers
