+++ src/media/ros2/ros2_reader.h
@@ -2,70 +2,31 @@
 // Copyright(c) 2025 RealSense, Inc. All Rights Reserved.
 
 #pragma once
-#include <rosbag2_storage/serialized_bag_message.hpp>
-#include <rosbag2_storage/topic_metadata.hpp>
-#include <rosbag2_storage_default_plugins/sqlite/sqlite_storage.hpp>
-
-#include "ros2_file_format.h"
+#include "ros2_reader_base.h"
 
 
 namespace librealsense
 {
     using namespace device_serializer;
 
-    class context;
-    class frame_source;
     class info_container;
     class options_interface;
     class options_container;
     class processing_block_interface;
     class recommended_proccesing_blocks_snapshot;
 
-    class ros2_reader : public reader
+    class ros2_reader : public ros2_reader_base
     {
     public:
         ros2_reader(const std::string& file, const std::shared_ptr<context> ctx);
-        device_snapshot query_device_description(const nanoseconds& time) override;
         std::shared_ptr<serialized_data> read_next_data() override;
-        void seek_to_time(const nanoseconds& seek_time) override;
-        std::vector<std::shared_ptr<serialized_data>> fetch_last_frames(const nanoseconds& seek_time) override;
-        nanoseconds query_duration() const override;
-        void reset() override;
-        virtual void enable_stream(const std::vector<device_serializer::stream_identifier>& stream_ids) override;
-        virtual void disable_stream(const std::vector<device_serializer::stream_identifier>& stream_ids) override;
-        const std::string& get_file_name() const override;
 
     private:
-        // We use a simple caching mechanism to have a lookahead functionality
-        // needed in some cases to tell what is the next message without missing it
-        bool has_next_cached() const;
-        std::shared_ptr<rosbag2_storage::SerializedBagMessage> read_next_cached();
-        std::shared_ptr<rosbag2_storage::SerializedBagMessage> peek_next_cached();
-
         static std::vector<std::string> split_string(const std::string& s, char delimiter);
         static std::string get_value(const std::map<std::string, std::string>& kv, const std::string& key);
         std::vector<std::string> filter_topics_by_regex(const std::regex& re) const;
         static std::map< std::string, std::string > parse_msg_payload(const std::shared_ptr<rosbag2_storage::SerializedBagMessage> msg);
         static void register_camera_infos(std::shared_ptr<info_container> infos, const std::map<std::string, std::string>& kv);
-
-        template<typename T>
-        static T deserialize_message(const std::shared_ptr<rosbag2_storage::SerializedBagMessage>& msg)
-        {
-            if (!msg || !msg->serialized_data || !msg->serialized_data->buffer || msg->serialized_data->buffer_length == 0)
-                throw std::runtime_error("Invalid message for deserialize_message, expected non-empty payload");
-
-            // Deserialize the message using Fast CDR
-            eprosima::fastcdr::FastBuffer fb(
-                reinterpret_cast<char*>(msg->serialized_data->buffer),
-                msg->serialized_data->buffer_length);
-            eprosima::fastcdr::Cdr cdr(fb, eprosima::fastcdr::Cdr::DEFAULT_ENDIAN, eprosima::fastcdr::Cdr::DDS_CDR);
-            cdr.read_encapsulation();
-            T data{};
-            data.deserialize(cdr);
-            return data;
-        }
-
-        nanoseconds get_file_duration();
 
         uint32_t read_file_version();
         bool try_read_stream_extrinsic(const stream_identifier& stream_id, uint32_t& group_id, rs2_extrinsics& extrinsic);
@@ -84,8 +45,10 @@ namespace librealsense
 
         std::shared_ptr<recommended_proccesing_blocks_snapshot> read_proccesing_blocks(device_serializer::sensor_identifier sensor_id,
             std::shared_ptr<options_interface> options);
-        device_snapshot read_device_description(const nanoseconds& time, bool reset = false);
-        void prepare_for_streaming();
+        device_snapshot read_device_description(const nanoseconds& time, bool reset = false) override;
+        std::vector<std::string> get_stream_topics() const override;
+        std::vector<std::string> get_option_topics() const override;
+        std::vector<std::string> get_notification_topics() const override;
 
         // Topic parsing helpers
         static bool is_stream_topic(const std::string& topic, stream_identifier& id);
@@ -104,10 +67,7 @@ namespace librealsense
 
         // Frame setup helpers
         void read_frame_metadata(frame_additional_data& additional_data);
-        void setup_frame(frame_interface* frame_ptr, const stream_identifier& sid) const;
-        
-        frame_holder alloc_and_move_frame(std::vector<uint8_t>&& data,
-            const stream_identifier& stream_id, frame_additional_data additional_data) const;
+        void setup_frame(frame_interface* frame_ptr, const stream_identifier& sid) const override;
 
         std::pair<rs2_option, std::shared_ptr<librealsense::option>> create_option(const std::shared_ptr<rosbag2_storage::SerializedBagMessage> msg);
         std::shared_ptr< serialized_frame > create_frame(const std::shared_ptr<rosbag2_storage::SerializedBagMessage> msg);
@@ -120,34 +80,8 @@ namespace librealsense
         notification create_notification(const std::shared_ptr<rosbag2_storage::SerializedBagMessage> msg) const;
         std::shared_ptr<options_container> read_sensor_options(device_serializer::sensor_identifier sensor_id);
 
-        std::shared_ptr< rosbag2_storage::storage_interfaces::ReadWriteInterface > _storage;
-
-        std::shared_ptr<metadata_parser_map>    m_metadata_parser_map;
-        device_snapshot                         m_initial_device_description;
-        nanoseconds                             m_total_duration;
-        std::string                             m_file_path;
-        std::shared_ptr<frame_source>           m_frame_source;
-        std::vector< rosbag2_storage::TopicMetadata > _topics_cache;
-        std::shared_ptr<context>                m_context;
         std::map<uint32_t, std::map<rs2_option, std::string>> m_read_options_descriptions;
 
-        // State management
-        bool _initialized = false;
-        std::set< stream_identifier > _enabled_streams;
-
-        // Cache to support fetch_last_frames logic
-        // Maps stream ID to the last frame data seen
-        std::map< stream_identifier, std::shared_ptr<serialized_data> > _last_frame_cache;
-
         std::map< stream_identifier, std::pair< uint32_t, rs2_extrinsics > > m_extrinsics_map;
-
-        static void decompress_if_needed(std::shared_ptr<rosbag2_storage::SerializedBagMessage>& msg);
-
-        std::shared_ptr<rosbag2_storage::SerializedBagMessage> _cached_message;
-        bool _cache_valid = false;  // true means _cached_message contains valid unconsumed data
-
-        // Filter topics for streaming - reapplied on reset() if set
-        std::vector<std::string> _streaming_filter_topics;
-
     };
 }
