+++ src/media/ros2/ros2_reader_base.cpp
@@ -1,0 +1,244 @@
+// License: Apache 2.0. See LICENSE file in root directory.
+// Copyright(c) 2026 RealSense, Inc. All Rights Reserved.
+
+#include "ros2_reader_base.h"
+
+#include <zstd.h>
+
+#include <src/context.h>
+#include <src/sensor.h>
+#include <src/stream.h>
+#include <src/core/extension.h>
+
+
+namespace librealsense
+{
+    ros2_reader_base::ros2_reader_base(const std::string& file, const std::shared_ptr<context> ctx)
+        : m_metadata_parser_map(md_constant_parser::create_metadata_parser_map())
+        , m_total_duration(0)
+        , m_file_path(file)
+        , m_context(ctx)
+    {
+        try
+        {
+            reset();
+            m_total_duration = get_file_duration();
+        }
+        catch (const std::exception& e)
+        {
+            throw io_exception(rsutils::string::from() << "Failed to create ros2 reader: " << e.what());
+        }
+    }
+
+    device_snapshot ros2_reader_base::query_device_description(const nanoseconds& time)
+    {
+        return read_device_description(time);
+    }
+
+    nanoseconds ros2_reader_base::query_duration() const
+    {
+        return m_total_duration;
+    }
+
+    nanoseconds ros2_reader_base::get_file_duration()
+    {
+        auto meta = _storage->get_metadata();
+        return nanoseconds(meta.duration.count());
+    }
+
+    void ros2_reader_base::reset()
+    {
+        _storage = std::make_shared< rosbag2_storage_plugins::SqliteStorage >();
+        _storage->open(m_file_path, rosbag2_storage::storage_interfaces::IOFlag::READ_ONLY);
+        m_frame_source = std::make_shared<frame_source>(32);
+        m_frame_source->init(m_metadata_parser_map);
+        _cache_valid = false;
+
+        // Reapply streaming filter if it was previously set
+        if (!_streaming_filter_topics.empty())
+        {
+            _storage->set_filter({ _streaming_filter_topics });
+        }
+    }
+
+    void ros2_reader_base::enable_stream(const std::vector<device_serializer::stream_identifier>& stream_ids)
+    {
+        for (const auto& id : stream_ids) _enabled_streams.insert(id);
+    }
+
+    void ros2_reader_base::disable_stream(const std::vector<device_serializer::stream_identifier>& stream_ids)
+    {
+        for (const auto& id : stream_ids) _enabled_streams.erase(id);
+    }
+
+    std::vector<std::shared_ptr<serialized_data>> ros2_reader_base::fetch_last_frames(const nanoseconds& seek_time)
+    {
+        std::vector<std::shared_ptr<serialized_data>> frames;
+        for (auto&& kv : _last_frame_cache)
+        {
+            // Filter by enabled streams
+            if (_enabled_streams.empty() || _enabled_streams.count(kv.first))
+            {
+                if (kv.second) frames.push_back(kv.second);
+            }
+        }
+        return frames;
+    }
+
+    const std::string& ros2_reader_base::get_file_name() const
+    {
+        return m_file_path;
+    }
+
+    bool ros2_reader_base::is_zstd_compressed(const uint8_t* src, size_t src_size)
+    {
+        return src_size >= 4 && src[0] == 0x28 && src[1] == 0xB5 && src[2] == 0x2F && src[3] == 0xFD;
+    }
+
+    void ros2_reader_base::decompress_if_needed(std::shared_ptr<rosbag2_storage::SerializedBagMessage>& msg)
+    {
+        if (!msg || !msg->serialized_data || !msg->serialized_data->buffer || msg->serialized_data->buffer_length == 0)
+            return;
+
+        auto src = msg->serialized_data->buffer;
+        auto src_size = msg->serialized_data->buffer_length;
+
+        if (!is_zstd_compressed(src, src_size))
+            return;
+
+        auto frame_content_size = ZSTD_getFrameContentSize(src, src_size);
+        if (frame_content_size == ZSTD_CONTENTSIZE_UNKNOWN || frame_content_size == ZSTD_CONTENTSIZE_ERROR)
+            throw std::runtime_error("Failed to determine decompressed size for zstd-compressed message");
+
+        // Guard against malformed frames claiming an absurd decompressed size — zstd's frame
+        // header is untrusted input and a malicious file could request a huge allocation.
+        constexpr size_t MAX_DECOMPRESSED_SIZE = 256 * 1024 * 1024; // 256 MB
+        if (frame_content_size > MAX_DECOMPRESSED_SIZE)
+            throw std::runtime_error(rsutils::string::from()
+                << "Zstd decompressed size " << frame_content_size << " exceeds safety limit");
+
+        auto decompressed_size = static_cast<size_t>(frame_content_size);
+
+        // We create a new buffer for the decompressed data each time. We could use
+        // a reusable member buffer like the writer does, but here the frame data may
+        // still be in use when the metadata is read next and overwrite it — allowing
+        // it to reallocate for simplicity for now.
+        auto out = create_buffer(decompressed_size);
+
+        auto result = ZSTD_decompress(out->buffer, out->buffer_capacity, src, src_size);
+        if (ZSTD_isError(result))
+            throw std::runtime_error(rsutils::string::from() << "Zstd decompression failed: " << ZSTD_getErrorName(result));
+
+        out->buffer_length = result;
+        msg->serialized_data = std::move(out);
+    }
+
+    void ros2_reader_base::prepare_for_streaming()
+    {
+        // Reopen storage to reset the filter, and apply relevant filters for streaming
+        _storage = std::make_shared< rosbag2_storage_plugins::SqliteStorage >();
+        _storage->open(m_file_path, rosbag2_storage::storage_interfaces::IOFlag::READ_ONLY);
+
+        auto stream_topics       = get_stream_topics();
+        auto option_topics       = get_option_topics();
+        auto notification_topics = get_notification_topics();
+
+        _streaming_filter_topics.clear();
+        _streaming_filter_topics.insert(_streaming_filter_topics.end(), stream_topics.begin(), stream_topics.end());
+        _streaming_filter_topics.insert(_streaming_filter_topics.end(), option_topics.begin(), option_topics.end());
+        _streaming_filter_topics.insert(_streaming_filter_topics.end(), notification_topics.begin(), notification_topics.end());
+
+        _storage->set_filter({ _streaming_filter_topics });
+    }
+
+    void ros2_reader_base::seek_to_time(const nanoseconds& seek_time)
+    {
+        // read all messages up to the requested time, updating the last frame is done inside the read function
+
+        if (seek_time > m_total_duration)
+        {
+            throw invalid_value_exception( rsutils::string::from()
+                                           << "Requested time is out of playback length. (Requested = "
+                                           << seek_time.count() << ", Duration = " << m_total_duration.count() << ")" );
+        }
+
+        reset();
+
+        auto msg = peek_next_cached();
+        while (msg && nanoseconds(static_cast<int64_t>(msg->time_stamp) - _first_timestamp_ns) < seek_time)
+        {
+            read_next_cached();
+            msg = peek_next_cached();
+        }
+    }
+
+    bool ros2_reader_base::has_next_cached() const
+    {
+        // If we have a valid cached message, we have next
+        if (_cache_valid)
+            return true;
+
+        return _storage->has_next();
+    }
+
+    std::shared_ptr<rosbag2_storage::SerializedBagMessage> ros2_reader_base::read_next_cached()
+    {
+        // If cache is valid, return cached message and mark as consumed
+        if (_cache_valid)
+        {
+            _cache_valid = false;
+            return _cached_message;
+        }
+
+        // Otherwise, read from storage and return immediately (no caching)
+        if (!_storage->has_next())
+            return nullptr;
+
+        auto msg = _storage->read_next();
+        decompress_if_needed(msg);
+        return msg;
+    }
+
+    std::shared_ptr<rosbag2_storage::SerializedBagMessage> ros2_reader_base::peek_next_cached()
+    {
+        // If cache is valid, return cached message without consuming
+        if (_cache_valid)
+            return _cached_message;
+
+        // Otherwise, read from storage and cache it
+        if (!_storage->has_next())
+            return nullptr;
+
+        _cached_message = _storage->read_next();
+        decompress_if_needed(_cached_message);
+        _cache_valid = true;
+        return _cached_message;
+    }
+
+    frame_holder ros2_reader_base::alloc_and_move_frame(std::vector<uint8_t>&& data,
+        const stream_identifier& stream_id,
+        frame_additional_data additional_data) const
+    {
+        auto frame_ext = frame_source::stream_to_frame_types(stream_id.stream_type);
+        frame_interface* frame = m_frame_source->alloc_frame(
+            { stream_id.stream_type, stream_id.stream_index, frame_ext },
+            data.size(),
+            std::move(additional_data),
+            true);
+
+        if (frame == nullptr)
+        {
+            LOG_WARNING("Failed to allocate new frame");
+            return frame_holder{};
+        }
+
+        // Move the deserialized data directly — avoids a full memcpy of image data
+        auto base_frame = static_cast<librealsense::frame*>(frame);
+        base_frame->data = std::move(data);
+
+        setup_frame(frame, stream_id);
+
+        return frame_holder{ frame };
+    }
+
+}
